refactor methods from program to ado and github
